+++ mastodon4j/src/main/java/com/sys1yagi/mastodon4j/api/entity/Tag.kt
@@ -10,5 +10,19 @@ class Tag(
         val name: String = "",
 
         @SerializedName("url")
-        val url: String = "") {
+        val url: String = "",
+
+        @SerializedName("history")
+        val history: List<History> = emptyList()
+) {
+                class History(
+                        @SerializedName("day")
+                        val day: Int = 0,
+
+                        @SerializedName("uses")
+                        val uses: Int = 0,
+
+                        @SerializedName("accounts")
+                        val accounts: Int = 0
+                )
 }
