--- mastodon4j/src/main/java/com/sys1yagi/mastodon4j/api/entity/Trend.kt
@@ -1,25 +1,0 @@
-package com.sys1yagi.mastodon4j.api.entity
-
-import com.google.gson.annotations.SerializedName
-
-/**
- * see more https://github.com/tootsuite/documentation/blob/master/Using-the-API/API.md#trend
- */
-class Trend(
-        @SerializedName("name")
-        val name: String = "",
-
-        @SerializedName("history")
-        val history: List<History> = emptyList()) {
-
-        class History(
-                @SerializedName("day")
-                val day: Int = 0,
-
-                @SerializedName("uses")
-                val uses: Int = 0,
-
-                @SerializedName("accounts")
-                val accounts: Int = 0
-        )
-}
